// src/auxiliary/interfaces/Weather.ts
import { WeatherForecastDay } from './WeatherForecastDay';

export interface Weather {
  location: {
    name: string;
    region: string;
    country: string;
  };
  current: {
    temp_c: number;
    condition: {
      text: string;
      icon: string;
    };
    humidity: number;
    wind_kph: number;
  };
  forecast: {
    forecastday: WeatherForecastDay[];
  };
}